requirement.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.1
django-filter==1.1.0
djangorestframework==3.9.4
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.2
pytz==2019.1
PyYAML==5.1
requests==2.22.0
sqlparse==0.3.0
stripe==2.28.1
uritemplate==3.0.0
urllib3==1.25.2
